+++ src/preload.ts
@@ -35,16 +35,18 @@ contextBridge.exposeInMainWorld('electronAPI', {
   downloadUpdate: () => ipcRenderer.invoke('updater:download-update'),
   installUpdate: () => ipcRenderer.invoke('updater:install-update'),
   getVersion: () => ipcRenderer.invoke('updater:get-version'),
-  
+
   // Update event listeners
   onUpdateAvailable: (callback: (info: any) => void) =>
     ipcRenderer.on('update-available', (event, info) => callback(info)),
   onUpdateNotAvailable: (callback: (info: any) => void) =>
     ipcRenderer.on('update-not-available', (event, info) => callback(info)),
-  onUpdateError: (callback: (error: string) => void) =>
-    ipcRenderer.on('update-error', (event, error) => callback(error)),
   onDownloadProgress: (callback: (progress: any) => void) =>
     ipcRenderer.on('download-progress', (event, progress) => callback(progress)),
   onUpdateDownloaded: (callback: (info: any) => void) =>
-    ipcRenderer.on('update-downloaded', (event, info) => callback(info))
+    ipcRenderer.on('update-downloaded', (event, info) => callback(info)),
+  // Fired once, just after launch, when this run is on a newer version than the
+  // previous run (i.e. a silent update just applied).
+  onAppUpdated: (callback: (version: string) => void) =>
+    ipcRenderer.on('app-updated', (event, version) => callback(version))
 });